packages/config/pyproject.toml
[project]
name = "monorepo-config"
version = "0.1.0"
description = "Configuration and environment utilities for the monorepo"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "BEMIND TECHNOLOGY CO., LTD.", email = "info@bemind.tech" }
]
keywords = ["config", "environment", "settings"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monorepo_config"]
